+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Changed
 - No match if word is in another one
-- Match word with camel case notation
+- Match word using camel case notation
+- Recommended abbrs only
 
 ## 0.0.1